kdbx3: fix length of transform_rounds field (#222)
The official KeePass client produces and expects a 64 bit field, not a
32 bit field, for this parameter.

Using the incorrect length happens to parse just fine. The correct
length is encoded into the header so there is no parsing misalignment.
And because the field is little endian, parsing the field itself as a 32
bit field even though it is actually a 64 bit field also works, provided
that the field value fits into 32 bits, which it typically does.

However, if we write header field back out, we write it with a 32 bit
length, and KeePass rejects this. We weren't writing out the header
ourselves previously because kdbx.header.data wasn't being removed (see
https://github.com/libkeepass/pykeepass/issues/219#issuecomment-748692045).
However if we start doing that, such as to fix issue #219, then this bug
is revealed.

The fix is trivial: we change the declaration to be a 64 bit field to
match the official client.