Implement framework for flexible 2FA
This adds support for using the hmac-secret FIDO extension to
contribute keying material for a KeePass 4 file.

It does this by storing an additional XML statekeeping blob in the
outer ("public") header. This blob is designed to hold a variety of
different authentication factors, such as passwords, key files, and
Yubikey challenge-response devices.