Export to html and email.py
Hello,

If useful, you can use it fully or partially.
I use this approach to quickly look at an entry from my mailbox or from the HTML file (that is generated) on the tablet or phone or hard disk.
I know that KeepassXC exists but still this code can be useful.

This Python script will export the entries to an HTML file on the hard drive.
Optionally, send it to your Gmail email address.

Warning: passwords in plain text. Use wisely.